+++ examples/Frogger/LilypadObject.cpp
@@ -1,7 +1,16 @@
+//////////////////////////////////////////////////////
+// Copyright 2002 Levi VanOort Released under MIT
+//				lvanoort@aol.com 
+//
+// LilypadObject.cpp
+// 4-5-2002
+//////////////////////////////////////////////////////
+
 #include <stdlib.h>
 #include <gl/glut.h>
 #include <iostream.h>
 #include "LilypadObject.h"
+
 
 extern int WindowSizeX, WindowSizeY;
 
@@ -16,69 +25,82 @@ LilypadObject::LilypadObject()
 	ObjectState = ALIVE;
 }
 
+
 LilypadObject::~LilypadObject()
 { 
 
 }
 
-void LilypadObject::SetGameCoordinates(int x, int y)
+
+void LilypadObject::SetGameCoordinates(float x, float y)
 {
 	GameX=x;
 	GameY=y;
 }
 
-void LilypadObject::SetSize(int x, int y)
+
+void LilypadObject::SetSize(float x, float y)
 {
 	ObjectSizeX = x;
 	ObjectSizeY = y;
 }
+
 
 void LilypadObject::SetState(State TheState)
 {
 	ObjectState = TheState;
 }
 
-void LilypadObject::SetVelocityY(int Vel)
+
+void LilypadObject::SetVelocityY(float Vel)
 {
 	VelocityY = Vel;
 }
 
-void LilypadObject::SetVelocityX(int Vel)
+
+void LilypadObject::SetVelocityX(float Vel)
 {
 	VelocityX = Vel;
 }
 
-void LilypadObject::GetGameCoordinates(int& x, int& y)
+
+void LilypadObject::GetGameCoordinates(float& x, float& y)
 {
 	x = GameX;
 	y = GameY;
 }
 
-void LilypadObject::GetSize(int& x, int& y)
+
+void LilypadObject::GetSize(float& x, float& y)
 {
 	x = ObjectSizeX;
 	y = ObjectSizeY;
 }
+
 
 void LilypadObject::GetState(State& TheState)
 {
 	TheState = ObjectState;
 }
 
-void LilypadObject::GetVelocityX(int& Vel)
+
+void LilypadObject::GetVelocityX(float& Vel)
 {
 	Vel = VelocityX;
 }
 
-void LilypadObject::GetVelocityY(int& Vel)
+
+void LilypadObject::GetVelocityY(float& Vel)
 {
 	Vel = VelocityX;
 }
+
 
 ObjectName LilypadObject::GetObjectName()
 {
 	return Name;
 }
+
 
 void LilypadObject::CollisionResponse(MyObject* CollisionObject)
 {
@@ -91,10 +113,9 @@ void LilypadObject::CollisionResponse(MyObject* CollisionObject)
 		ObjectState = COLLISION;
 		break;
 	}
-
 }
 
-bool LilypadObject::Moved(int& x, int& y)
+bool LilypadObject::Moved(float& x, float& y)
 {
 	return false;
 }
@@ -103,6 +124,7 @@ void LilypadObject::Update()
 {	
 	NULL;
 }
+
 		 
 void LilypadObject::Draw()
 { 
@@ -111,9 +133,11 @@ void LilypadObject::Draw()
 	case ALIVE:
 		glColor3f(0.0,0.0,1.0);
 		break;
+
 	case COLLISION:
 		glColor3f(1.0,1.0,1.0);
 		break;
 	}
 	glRectf(float(GameX),float(GameY), float(GameX+ObjectSizeX), float(GameY+ObjectSizeY)); 
 }
+
